Refactor Flask routes to separate module
